requirements.txt
attrs==21.4.0
bcrypt==3.2.2
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
channel-app==0.0.99
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.1
Flask-Authorize==0.2.6
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.1
fonttools==4.31.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
lightgbm==3.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
passlib==1.7.4
Pillow==9.1.0
plotly==5.7.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.37
tenacity==8.0.1
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0